Move Type/Typed into Path.Value.
