+++ src/Path/Scope.hs
@@ -6,7 +6,6 @@ import Data.Coerce
 import qualified Data.Map as Map
 import Path.Name
 import Path.Pretty
-import Path.Type
 import Path.Value
 
 data Entry
